[refactor] Merge 해결 및 폴더명 변경
