+++ client/src/components/Ingredients/FontSizeMenu/style.ts
