+++ client/src/components/Meal/FormatButtons/useFormatButtons.ts
@@ -5,6 +5,7 @@ import useToggle from '@hooks/useToggle';
 const useFormatButtons = () => {
   const dispatch = useDispatch();
   const [sizeMenu, toggleSizeMenu] = useToggle(false);
+  const [colorMenu, toggleColorMenu] = useToggle(false);
 
   const alignHandler = (direction: string): void => {
     dispatch(changeTextAlign(direction));
@@ -16,6 +17,8 @@ const useFormatButtons = () => {
   return {
     sizeMenu,
     toggleSizeMenu,
+    colorMenu,
+    toggleColorMenu,
     alignHandler,
     clearHandler,
   };
