+++ client/src/contexts/latex/types.ts
@@ -25,7 +25,7 @@ export type LatexAction =
   | ReturnType<typeof changeTab>
   | ReturnType<typeof removeTab>
   | ReturnType<typeof changeFontSize>
-  | ReturnType<typeof changeFontColor>;
+  | ReturnType<typeof changeFontColor>
   | ReturnType<typeof changeTextAlign>;
 
 export interface LatexState {
