+++ client/src/hooks/useToggle.ts
@@ -1,13 +1,16 @@
 import { useState } from 'react';
 
-const useToggle = (initialValue: boolean): [boolean, () => void] => {
+const useToggle = (
+  initialValue: boolean
+): [boolean, () => void, React.Dispatch<React.SetStateAction<boolean>>] => {
   const [value, setValue] = useState(initialValue);
 
   const onToggle = () => {
+    // eslint-disable-next-line no-shadow
     setValue((value) => !value);
   };
 
-  return [value, onToggle];
+  return [value, onToggle, setValue];
 };
 
 export default useToggle;
